pyproject.toml
[project]
name = "hadoop-practice"
version = "0.1.0"
description = "Comprehensive Hadoop ecosystem learning project with HDFS, MapReduce, Hive, and Spark"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyspark>=3.0.0",
    "delta-spark>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
]
